refactor(frontend): adopt mutation hooks, split god components
- migrate imperative handleApiCall write paths to useApiMutation across the
  api-key/webhook/wallet/agent/x402 dialogs and pages (reads left as-is)
- split SetupWelcome (2108 lines) into per-screen components + wallet-
  generation/payment-source hooks
- split WalletDetailsDialog (2087 lines) into feature sections +
  useTokenBalances/useLowBalanceRules hooks; every file now <=750 lines
