+++ frontend/src/components/api-keys/UpdateApiKeyDialog.tsx
@@ -17,8 +17,8 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/components/ui/select';
-import { PatchApiKeyResponse } from '@/lib/api/generated/types.gen';
-import { handleApiCall } from '@/lib/utils';
+import { PatchApiKeyData, PatchApiKeyResponse } from '@/lib/api/generated/types.gen';
+import { useApiMutation } from '@/lib/hooks/useApiMutation';
 import { useForm, Controller, useWatch } from 'react-hook-form';
 import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
@@ -102,10 +102,16 @@ function getPermissionLabel(_canRead: boolean, canPay: boolean, canAdmin: boolea
 }
 
 export function UpdateApiKeyDialog({ open, onClose, onSuccess, apiKey }: UpdateApiKeyDialogProps) {
-  const [isLoading, setIsLoading] = useState(false);
   const [showToken, setShowToken] = useState(false);
   const tokenInputRef = useRef<HTMLInputElement | null>(null);
   const { apiClient, network } = useAppContext();
+
+  const updateApiKey = useApiMutation({
+    mutationFn: (body: NonNullable<PatchApiKeyData['body']>) =>
+      patchApiKey({ client: apiClient, body }),
+    errorMessage: 'Failed to update API key',
+  });
+  const isLoading = updateApiKey.isPending;
   const { paymentSources } = usePaymentSourceExtendedAll();
   const { wallets: managedWallets } = useAllWallets(open);
   // A key's NetworkLimit can span both Cardano networks, so offer EVM chains from every
@@ -159,77 +165,63 @@ export function UpdateApiKeyDialog({ open, onClose, onSuccess, apiKey }: UpdateA
   });
 
   const onSubmit = async (data: UpdateApiKeyFormValues) => {
-    setIsLoading(true);
-    try {
-      const usageCredits: Array<{ unit: string; amount: string }> = [];
-      if (data.credits.lovelace) {
-        usageCredits.push({
-          unit: 'lovelace',
-          amount: convertDecimalToBaseUnits(data.credits.lovelace),
-        });
-      }
-      if (data.credits.usdcx) {
-        usageCredits.push({
-          unit: getActiveStablecoinConfig(network).fullAssetId,
-          amount: convertDecimalToBaseUnits(data.credits.usdcx),
-        });
-      }
-
-      const walletScopeChanged =
-        data.walletScopeEnabled !== apiKey.walletScopeEnabled ||
-        JSON.stringify([...data.walletScopeIds].sort()) !==
-          JSON.stringify([...apiKey.WalletScopes.map((ws) => ws.hotWalletId)].sort());
-
-      const initialEvmChains = apiKey.ChainIdLimit.filter((chainId) =>
-        chainId.startsWith('eip155:'),
-      );
-      const evmChainsChanged =
-        JSON.stringify([...data.evmChains].sort()) !== JSON.stringify([...initialEvmChains].sort());
-
-      await handleApiCall(
-        () =>
-          patchApiKey({
-            client: apiClient,
-            body: {
-              id: apiKey.id,
-              ...(data.newToken && { token: data.newToken }),
-              ...(data.status !== apiKey.status && { status: data.status }),
-              ...(usageCredits.length > 0 && {
-                UsageCreditsToAddOrRemove: usageCredits,
-              }),
-              ...(walletScopeChanged && {
-                walletScopeEnabled: data.walletScopeEnabled,
-                WalletScopeHotWalletIds: data.walletScopeEnabled ? data.walletScopeIds : [],
-              }),
-              ...(apiKey.canPay &&
-                !apiKey.canAdmin &&
-                evmChainsChanged && {
-                  // ChainIdLimit replaces the whole networkLimit on update; preserve
-                  // the existing Cardano grants and only rewrite the EVM ones.
-                  ChainIdLimit: [
-                    ...apiKey.ChainIdLimit.filter((chainId) => !chainId.startsWith('eip155:')),
-                    ...data.evmChains,
-                  ],
-                }),
-            },
-          }),
-        {
-          onSuccess: (response) => {
-            const responseData = response?.data as PatchApiKeyResponse;
-            if (!responseData?.data?.id) {
-              toast.error('Failed to update API key: Invalid response from server');
-              return;
-            }
-            toast.success('API key updated successfully');
-            onSuccess();
-            handleClose();
-          },
-          errorMessage: 'Failed to update API key',
-        },
-      );
-    } finally {
-      setIsLoading(false);
+    const usageCredits: Array<{ unit: string; amount: string }> = [];
+    if (data.credits.lovelace) {
+      usageCredits.push({
+        unit: 'lovelace',
+        amount: convertDecimalToBaseUnits(data.credits.lovelace),
+      });
     }
+    if (data.credits.usdcx) {
+      usageCredits.push({
+        unit: getActiveStablecoinConfig(network).fullAssetId,
+        amount: convertDecimalToBaseUnits(data.credits.usdcx),
+      });
+    }
+
+    const walletScopeChanged =
+      data.walletScopeEnabled !== apiKey.walletScopeEnabled ||
+      JSON.stringify([...data.walletScopeIds].sort()) !==
+        JSON.stringify([...apiKey.WalletScopes.map((ws) => ws.hotWalletId)].sort());
+
+    const initialEvmChains = apiKey.ChainIdLimit.filter((chainId) => chainId.startsWith('eip155:'));
+    const evmChainsChanged =
+      JSON.stringify([...data.evmChains].sort()) !== JSON.stringify([...initialEvmChains].sort());
+
+    const response = await updateApiKey
+      .mutateAsync({
+        id: apiKey.id,
+        ...(data.newToken && { token: data.newToken }),
+        ...(data.status !== apiKey.status && { status: data.status }),
+        ...(usageCredits.length > 0 && {
+          UsageCreditsToAddOrRemove: usageCredits,
+        }),
+        ...(walletScopeChanged && {
+          walletScopeEnabled: data.walletScopeEnabled,
+          WalletScopeHotWalletIds: data.walletScopeEnabled ? data.walletScopeIds : [],
+        }),
+        ...(apiKey.canPay &&
+          !apiKey.canAdmin &&
+          evmChainsChanged && {
+            // ChainIdLimit replaces the whole networkLimit on update; preserve
+            // the existing Cardano grants and only rewrite the EVM ones.
+            ChainIdLimit: [
+              ...apiKey.ChainIdLimit.filter((chainId) => !chainId.startsWith('eip155:')),
+              ...data.evmChains,
+            ],
+          }),
+      })
+      .catch(() => null);
+    if (!response) return;
+
+    const responseData = response?.data as PatchApiKeyResponse;
+    if (!responseData?.data?.id) {
+      toast.error('Failed to update API key: Invalid response from server');
+      return;
+    }
+    toast.success('API key updated successfully');
+    onSuccess();
+    handleClose();
   };
 
   const handleClose = () => {
